ext-diet WP4: EnvoyExt read/introspection handlers move to envoy_read module DAT
26 read handlers + sole-caller helpers move behind delegating stubs;
EnvoyExt 7,403 -> 6,240 lines. Gate caught the run's first real defect:
_get_docs is WORKER-thread-served and mod.* is unavailable off the main
thread, so the docs family lives on the facade (rule now hard-coded in
the plan for every remaining WP). 3-reviewer panel confirmed: 30/30
bodies AST-faithful, thread-boundary closure clean (zero worker-side
mod.*/TD access), all 53 tool chains audited end-to-end OK. Hardening:
envoyVersion lookup falls back to 'unknown' instead of failing
get_td_info; doubled @staticmethod scar fixed; stale module imports and
docs comments removed.
