+++ dev/embody/externalizations.tsv
@@ -101,7 +101,7 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters2	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters2.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters1	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters1.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
-/embody/unit_tests/test_sandbox	base	tdn	embody/unit_tests/test_sandbox.tdn	2026-07-06 04:59:03 UTC		606	099.2025.32820	400	0	0.3000,0.5000,0.9000
+/embody/unit_tests/test_sandbox	base	tdn	embody/unit_tests/test_sandbox.tdn	2026-07-06 05:28:27 UTC		606	099.2025.32820	400	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_rename_move_lifecycle	text	py	embody/unit_tests/test_rename_move_lifecycle.py	2026-06-20 20:36:03 UTC				200	-600	0.9000,0.3000,0.4000
 /embody/unit_tests/test_path_utils	text	py	embody/unit_tests/test_path_utils.py	2026-06-20 20:36:03 UTC				200	-800	0.9000,0.3000,0.4000
 /embody/unit_tests/test_param_tracker	text	py	embody/unit_tests/test_param_tracker.py	2026-06-20 20:36:03 UTC				400	-800	0.9000,0.3000,0.4000
